- Move checkout screen from /payment to /checkout - Remove unused PurchaseItem.tsx
