+++ client/src/renderer/ui/components/pos/checkout/Main.tsx
